alterando de double para int

+++ Calculadora/Calculadora/Domain/Operations.cs
@@ -2,11 +2,11 @@
 {
     public class Operations
     {
-        public double Sum(double n1, double n2) => n1 + n2;
-        public double Subtraction(double n1, double n2) => n1 - n2;
+        public int Sum(int n1, int n2) => n1 + n2;
+        public int Subtraction(int n1, int n2) => n1 - n2;
 
-        public double Multiplication(double n1, double n2) => n1 * n2;
+        public int Multiplication(int n1, int n2) => n1 * n2;
 
-        public double Division(double n1, double n2) => n1 / n2;
+        public int Division(int n1, int n2) => n1 / n2;
     }
 }
